+++ test/instream/series/hydrator_test.exs
@@ -16,17 +16,29 @@ defmodule Instream.Series.HydratorTest do
   test "hydrating from map", %{test: test} do
     val_field = System.unique_integer()
     val_tag = Atom.to_string(test)
-    val_timestamp = System.unique_integer([:positive])
 
-    assert %TestSeries{
-             fields: %{value: ^val_field},
-             tags: %{foo: ^val_tag},
-             timestamp: ^val_timestamp
-           } =
+    # `DateTime` only supports :microsecond level precisions
+    val_timestamp = System.unique_integer([:positive]) * 10_000
+    val_datetime = val_timestamp |> DateTime.from_unix!(:nanosecond) |> DateTime.to_iso8601()
+
+    expected = %TestSeries{
+      fields: %TestSeries.Fields{value: val_field},
+      tags: %TestSeries.Tags{foo: val_tag},
+      timestamp: val_timestamp
+    }
+
+    assert ^expected =
              TestSeries.from_map(%{
                foo: val_tag,
                value: val_field,
                timestamp: val_timestamp
+             })
+
+    assert ^expected =
+             TestSeries.from_map(%{
+               foo: val_tag,
+               value: val_field,
+               timestamp: val_datetime
              })
   end
 
@@ -50,21 +62,27 @@ defmodule Instream.Series.HydratorTest do
     val_field_1 = System.unique_integer()
     val_field_2 = System.unique_integer()
     val_tag = Atom.to_string(test)
-    val_timestamp_1 = System.unique_integer([:positive])
-    val_timestamp_2 = System.unique_integer([:positive])
 
-    assert [
-             %TestSeries{
-               fields: %{value: ^val_field_1},
-               tags: %{foo: ^val_tag},
-               timestamp: ^val_timestamp_1
-             },
-             %TestSeries{
-               fields: %{value: ^val_field_2},
-               tags: %{foo: ^val_tag},
-               timestamp: ^val_timestamp_2
-             }
-           ] =
+    # `DateTime` only supports :microsecond level precisions
+    val_timestamp_1 = System.unique_integer([:positive]) * 10_000
+    val_timestamp_2 = System.unique_integer([:positive]) * 10_000
+    val_datetime_1 = val_timestamp_1 |> DateTime.from_unix!(:nanosecond) |> DateTime.to_iso8601()
+    val_datetime_2 = val_timestamp_2 |> DateTime.from_unix!(:nanosecond) |> DateTime.to_iso8601()
+
+    expected = [
+      %TestSeries{
+        fields: %TestSeries.Fields{value: val_field_1},
+        tags: %TestSeries.Tags{foo: val_tag},
+        timestamp: val_timestamp_1
+      },
+      %TestSeries{
+        fields: %TestSeries.Fields{value: val_field_2},
+        tags: %TestSeries.Tags{foo: val_tag},
+        timestamp: val_timestamp_2
+      }
+    ]
+
+    assert ^expected =
              TestSeries.from_result(%{
                results: [
                  %{
@@ -74,6 +92,22 @@ defmodule Instream.Series.HydratorTest do
                        name: "hydrator_test",
                        tags: %{foo: val_tag},
                        values: [[val_timestamp_1, val_field_1], [val_timestamp_2, val_field_2]]
+                     }
+                   ]
+                 }
+               ]
+             })
+
+    assert ^expected =
+             TestSeries.from_result(%{
+               results: [
+                 %{
+                   series: [
+                     %{
+                       columns: ["time", "value"],
+                       name: "hydrator_test",
+                       tags: %{foo: val_tag},
+                       values: [[val_datetime_1, val_field_1], [val_datetime_2, val_field_2]]
                      }
                    ]
                  }